factor out some functions to sc_utils module
